rename module x80 to kr580vm1
